requirements.txt
gradio
transformers
torch
soundfile
pymupdf
pytesseract
Pillow
datasets
num2words
sentence-transformers
sentencepiece